# golang-mongodb-kafka/api/docs/swagger.yaml
definitions:
  models.BadRequestError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.CreateProductRequest:
    properties:
      id:
        type: string
      name:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.Photos'
        type: array
      price:
        type: integer
    type: object
  models.CreateResponse:
    properties:
      id:
        type: string
    type: object
  models.GetAllProductsResponse:
    properties:
      count:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
  models.GetProductResponse:
    properties:
      product:
        $ref: '#/definitions/models.Product'
    type: object
  models.InternalServerError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.Photos:
    properties:
      id:
        type: string
    type: object
  models.Product:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.Photos'
        type: array
      price:
        type: integer
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/product:
    get:
      consumes:
      - application/json
      description: API for getting all Products
      parameters:
      - description: name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllProductsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerError'
      summary: Get All Products
      tags:
      - product
    post:
      consumes:
      - application/json
      description: API for creating product
      parameters:
      - description: product
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerError'
      summary: Create product
      tags:
      - product
  /v1/product/{product_id}:
    get:
      consumes:
      - application/json
      description: API for getting a product
      parameters:
      - description: product_id
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerError'
      summary: Get Product
      tags:
      - product
    put:
      consumes:
      - application/json
      description: API for creating product
      parameters:
      - description: product_id
        in: path
        name: product_id
        required: true
        type: string
      - description: product
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/models.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BadRequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.InternalServerError'
      summary: Update product
      tags:
      - product
swagger: "2.0"